+++ src/modules/launcher/Wox.Plugin/Common/DefaultBrowserInfo.cs
