+++ src/modules/launcher/Wox.Plugin/Common/ShellCommand.cs
